requirements.txt
cffi==1.17.1
cryptography==43.0.3
oracledb==2.5.0
paho-mqtt==2.1.0
pycparser==2.22
python-dotenv==1.0.1
